+++ util/jenkins/ansible-provision.sh
@@ -180,16 +180,16 @@ if [[ -z $registrar ]]; then
   registrar="false"
 fi
 
-if [[ -z $registrar_version ]]; then
-  registrar_version="master"
+if [[ -z $REGISTRAR_VERSION ]]; then
+  REGISTRAR_VERSION="master"
 fi
 
 if [[ -z $learner_portal ]]; then
   learner_portal="false"
 fi
 
-if [[ -z $learner_portal_version ]]; then
-  learner_portal_version="master"
+if [[ -z $LEARNER_PORTAL_VERSION ]]; then
+  LEARNER_PORTAL_VERSION="master"
 fi
 
 if [[ $registrar == 'true' ]]; then
@@ -206,17 +206,17 @@ ssh-keygen -f "/var/lib/jenkins/.ssh/known_hosts" -R "$deploy_host"
 cd playbooks
 
 cat << EOF > $extra_vars_file
-edx_platform_version: $edxapp_version
-forum_version: $forum_version
+EDX_PLATFORM_VERSION: $edxapp_version
+FORUM_VERSION: $FORUM_VERSION
 notifier_version: $notifier_version
 XQUEUE_VERSION: $xqueue_version
-certs_version: $certs_version
-configuration_version: $configuration_version
-demo_version: $demo_version
+CERTS_VERSION: $CERTS_VERSION
+CONFIGURATION_VERSION: $CONFIGURATION_VERSION
+DEMO_VERSION: $DEMO_VERSION
 THEMES_VERSION: $themes_version
-registrar_version: $registrar_version
-learner_portal_version: $learner_portal_version
-program_manager_version: $program_manager_version
+REGISTRAR_VERSION: $REGISTRAR_VERSION
+LEARNER_PORTAL_VERSION: $LEARNER_PORTAL_VERSION
+PROGRAM_MANAGER_VERSION: $PROGRAM_MANAGER_VERSION
 
 edx_ansible_source_repo: ${configuration_source_repo}
 edx_platform_repo: ${edx_platform_repo}
@@ -242,18 +242,18 @@ ANALYTICS_API_VERSION: $analytics_api_version
 
 REGISTRAR_NGINX_PORT: 80
 REGISTRAR_SSL_NGINX_PORT: 443
-REGISTRAR_VERSION: $registrar_version
+REGISTRAR_VERSION: $REGISTRAR_VERSION
 REGISTRAR_ENABLED: $registrar
 
 LEARNER_PORTAL_NGINX_PORT: 80
 LEARNER_PORTAL_SSL_NGINX_PORT: 443
-LEARNER_PORTAL_VERSION: $learner_portal_version
+LEARNER_PORTAL_VERSION: $LEARNER_PORTAL_VERSION
 LEARNER_PORTAL_ENABLED: $learner_portal
 LEARNER_PORTAL_SANDBOX_BUILD: True
 
 PROGRAM_MANAGER_NGINX_PORT: 80
 PROGRAM_MANAGER_SSL_NGINX_PORT: 443
-PROGRAM_MANAGER_VERSION: $program_manager_version
+PROGRAM_MANAGER_VERSION: $PROGRAM_MANAGER_VERSION
 PROGRAM_MANAGER_ENABLED: $program_manager
 PROGRAM_MANAGER_SANDBOX_BUILD: True